Consolidate server files
