+++ openhtf/output/servers/dashboard_server.py
@@ -18,9 +18,9 @@ import sockjs.tornado
 import tornado.web
 
 from openhtf.output.servers import station_server
-from openhtf.output.web_gui_server import pub_sub
-from openhtf.output.web_gui_server import web_gui_server
-from openhtf.output.web_gui_server import web_launcher
+from openhtf.output.servers import pub_sub
+from openhtf.output.servers import web_gui_server
+from openhtf.output.web_gui import web_launcher
 from openhtf.util import data
 from openhtf.util import multicast
