+++ openhtf/output/servers/station_server.py
@@ -24,8 +24,8 @@ import tornado.web
 
 import openhtf
 from openhtf.output.callbacks import mfg_inspector
-from openhtf.output.web_gui_server import pub_sub
-from openhtf.output.web_gui_server import web_gui_server
+from openhtf.output.servers import pub_sub
+from openhtf.output.servers import web_gui_server
 from openhtf.util import conf
 from openhtf.util import data
 from openhtf.util import functions
